Version 2.4.4
+ Gave all machines the IItemHandler capability
> Improved internal infusion recipe handling
> Made the Demon Warding potion use a Mundane potion as a base rather
than a water bottle
- Removed oredict entries for the Demon Warding potions, as they are no
longer needed
> Cleaned up imports and such

+++ src/main/java/sblectric/lightningcraft/blocks/BlockLightningInfuser.java
@@ -108,9 +108,9 @@ public class BlockLightningInfuser extends BlockContainerLC implements IFurnace 
 	@Override
 	public void randomDisplayTick(IBlockState state, World world, BlockPos pos, Random rand) {
 		if(isBurning(state, world, pos)) {
-			double d0 = (double)pos.getX() + 0.5D;
-			double d1 = (double)pos.getY() + rand.nextDouble() * 6.0D / 16.0D;
-			double d2 = (double)pos.getZ() + 0.5D;
+			double d0 = pos.getX() + 0.5D;
+			double d1 = pos.getY() + rand.nextDouble() * 6.0D / 16.0D;
+			double d2 = pos.getZ() + 0.5D;
 			double d3 = 0.52D;
 			double d4 = rand.nextDouble() * 0.6D - 0.3D;
