+++ src/main/java/sblectric/lightningcraft/entities/EntityUnderworldGhast.java
@@ -145,14 +145,14 @@ public class EntityUnderworldGhast extends EntityGhast {
 					double d1 = 4.0D;
 					Vec3d vec3d = this.parentEntity.getLook(1.0F);
 					double d2 = entitylivingbase.posX - (this.parentEntity.posX + vec3d.xCoord * d1);
-					double d3 = entitylivingbase.getEntityBoundingBox().minY + (double)(entitylivingbase.height / 2.0F) - 
-							(0.5D + this.parentEntity.posY + (double)(this.parentEntity.height / 2.0F));
+					double d3 = entitylivingbase.getEntityBoundingBox().minY + entitylivingbase.height / 2.0F - 
+							(0.5D + this.parentEntity.posY + this.parentEntity.height / 2.0F);
 					double d4 = entitylivingbase.posZ - (this.parentEntity.posZ + vec3d.zCoord * d1);
 					world.playEvent((EntityPlayer)null, 1016, new BlockPos(this.parentEntity), 0);
 					EntityLargeFireball entitylargefireball = new EntityLargeFireball(world, this.parentEntity, d2, d3, d4);
 					entitylargefireball.explosionPower = this.parentEntity.getFireballStrength();
 					entitylargefireball.posX = this.parentEntity.posX + vec3d.xCoord * d1;
-					entitylargefireball.posY = this.parentEntity.posY + (double)(this.parentEntity.height / 2.0F) + 0.5D;
+					entitylargefireball.posY = this.parentEntity.posY + this.parentEntity.height / 2.0F + 0.5D;
 					entitylargefireball.posZ = this.parentEntity.posZ + vec3d.zCoord * d1;
 					world.spawnEntityInWorld(entitylargefireball);
 					this.attackTimer = -30;
